requirements.txt
Django==4.2.7
mysqlclient==2.2.0
qrcode[pil]==7.4.2
segno==1.5.2
Pillow==10.1.0
aiogram==3.13.1
django-tables2==2.6.0
reportlab==4.0.7
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
requests==2.31.0
gunicorn==21.2.0

